pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kalshi-scanner"
version = "0.1.0"
description = "Kalshi crypto 15-min fair-value scanner — Phase 1 feasibility research."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kalshi_python_sync",
    "websockets>=12",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "psycopg2-binary>=2.9",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "lib/trading_platform/src", "scripts"]
